refactor(HTML encoding): Aggregate into `generics.rs`
